+++ src/Kiota.Builder/Refiners/JavaRefiner.cs
@@ -28,7 +28,7 @@ public class JavaRefiner : CommonLanguageRefiner, ILanguageRefiner
             AddRawUrlConstructorOverload(generatedCode);
             CorrectCoreType(generatedCode, CorrectMethodType, CorrectPropertyType, CorrectImplements);
             cancellationToken.ThrowIfCancellationRequested();
-            ReplaceBinaryByNativeType(generatedCode, "InputStream", "java.io", true);
+            ReplaceBinaryByNativeType(generatedCode, "InputStream", "java.io", true, true);
             AddGetterAndSetterMethods(generatedCode,
                 new() {
                     CodePropertyKind.Custom,
@@ -81,7 +81,40 @@ public class JavaRefiner : CommonLanguageRefiner, ILanguageRefiner
                 "ParseNode",
                 addUsings: true
             );
+            RemoveHandlerFromRequestBuilder(generatedCode);
+            SplitLongDiscriminatorMethods(generatedCode);
         }, cancellationToken);
+    }
+    private static readonly int maxDiscriminatorLength = 500;
+    private static void SplitLongDiscriminatorMethods(CodeElement currentElement) {
+        if (currentElement is CodeMethod currentMethod &&
+            !currentMethod.IsOverload &&
+            currentMethod.IsOfKind(CodeMethodKind.Factory) &&
+            currentMethod.Parent is CodeClass parentClass &&
+            parentClass.IsOfKind(CodeClassKind.Model) &&
+            parentClass.DiscriminatorInformation.HasBasicDiscriminatorInformation &&
+            parentClass.DiscriminatorInformation.DiscriminatorMappings.Count() > maxDiscriminatorLength) {
+                var discriminatorsCount = parentClass.DiscriminatorInformation.DiscriminatorMappings.Count();
+                for(var currentDiscriminatorPageIndex = 0; currentDiscriminatorPageIndex * maxDiscriminatorLength < discriminatorsCount; currentDiscriminatorPageIndex++) {
+                    var newMethod = currentMethod.Clone() as CodeMethod;
+                    newMethod.Name = $"{currentMethod.Name}_{currentDiscriminatorPageIndex}";
+                    newMethod.OriginalMethod = currentMethod;
+                    newMethod.Access = AccessModifier.Private;
+                    newMethod.RemoveParametersByKind(CodeParameterKind.ParseNode);
+                    newMethod.AddParameter(new CodeParameter {
+                        Type = new CodeType {
+                            Name = "String",
+                            IsNullable = true,
+                            IsExternal = true
+                        },
+                        Optional = false,
+                        Description = "Discriminator value from the payload",
+                        Name = "discriminatorValue"
+                    });
+                    parentClass.AddMethod(newMethod);
+                }
+            }
+        CrawlTree(currentElement, SplitLongDiscriminatorMethods);
     }
     private static void SetSetterParametersToNullable(CodeElement currentElement, params Tuple<CodeMethodKind, CodePropertyKind>[] accessorPairs) {
         if(currentElement is CodeMethod method &&
@@ -114,8 +147,6 @@ public class JavaRefiner : CommonLanguageRefiner, ILanguageRefiner
             "java.net", "URISyntaxException"),
         new (static x => x is CodeMethod method && method.IsOfKind(CodeMethodKind.RequestGenerator),
             "java.util", "Collection", "Map"),
-        new (static x => x is CodeMethod method && method.IsOfKind(CodeMethodKind.RequestExecutor),
-            "com.microsoft.kiota", "ResponseHandler"),
         new (static x => x is CodeClass @class && @class.IsOfKind(CodeClassKind.Model),
             "com.microsoft.kiota.serialization", "Parsable"),
         new (static x => x is CodeClass @class && @class.IsOfKind(CodeClassKind.Model) && @class.Properties.Any(x => x.IsOfKind(CodePropertyKind.AdditionalData)),
@@ -186,11 +217,7 @@ public class JavaRefiner : CommonLanguageRefiner, ILanguageRefiner
         block.Implements.Where(x => "IAdditionalDataHolder".Equals(x.Name, StringComparison.OrdinalIgnoreCase)).ToList().ForEach(x => x.Name = x.Name[1..]); // skipping the I
     }
     private static void CorrectMethodType(CodeMethod currentMethod) {
-        if(currentMethod.IsOfKind(CodeMethodKind.RequestExecutor, CodeMethodKind.RequestGenerator)) {
-            if(currentMethod.IsOfKind(CodeMethodKind.RequestExecutor))
-                currentMethod.Parameters.Where(x => x.IsOfKind(CodeParameterKind.ResponseHandler) && x.Type.Name.StartsWith("i", StringComparison.OrdinalIgnoreCase)).ToList().ForEach(x => x.Type.Name = x.Type.Name[1..]);
-        }
-        else if(currentMethod.IsOfKind(CodeMethodKind.Serializer))
+        if(currentMethod.IsOfKind(CodeMethodKind.Serializer))
             currentMethod.Parameters.Where(x => x.IsOfKind(CodeParameterKind.Serializer)).ToList().ForEach(x => {
                 x.Optional = false;
                 x.Type.IsNullable = true;
@@ -255,9 +282,8 @@ public class JavaRefiner : CommonLanguageRefiner, ILanguageRefiner
             if(codeMethods.Any()) {
                 var originalExecutorMethods = codeMethods.Where(x => x.IsOfKind(CodeMethodKind.RequestExecutor));
                 var executorMethodsToAdd = originalExecutorMethods
-                                    .Select(x => GetMethodClone(x, CodeParameterKind.ResponseHandler))
                                     .Union(originalExecutorMethods
-                                            .Select(x => GetMethodClone(x, CodeParameterKind.RequestConfiguration, CodeParameterKind.ResponseHandler)))
+                                            .Select(x => GetMethodClone(x, CodeParameterKind.RequestConfiguration)))
                                     .Where(x => x != null);
                 var originalGeneratorMethods = codeMethods.Where(x => x.IsOfKind(CodeMethodKind.RequestGenerator));
                 var generatorMethodsToAdd = originalGeneratorMethods