+++ src/Kiota.Builder/SearchProviders/APIsGuru/APIsGuruSearchProvider.cs
@@ -42,7 +42,7 @@ public class APIsGuruSearchProvider : ISearchProvider
                                 .Select(x => x.Value.versions.TryGetValue(GetVersionKey(singleCandidate, version, x), out var versionInfo) ? (x.Key, versionInfo, x.Value.versions.Keys.ToList()) : (x.Key, default, default))
                                 .Where(static x => x.versionInfo is not null)
                                 .ToDictionary(static x => x.Key,
-                                            static x => new SearchResult(x.versionInfo.info?.title, x.versionInfo.info?.description, x.versionInfo.info?.contact?.url, x.versionInfo.info?.origin?.FirstOrDefault()?.url, x.Item3),
+                                            static x => new SearchResult(x.versionInfo.info?.title, x.versionInfo.info?.description, x.versionInfo.info?.contact?.url, x.versionInfo.swaggerUrl, x.Item3),
                                             StringComparer.OrdinalIgnoreCase);
         return results;
     }